+++ backend/Controllers/Users/AuthenticateController.cs
@@ -65,69 +65,5 @@ namespace BackendAPI.Controllers
             var encodedJwt = handler.WriteToken(token);
             return encodedJwt;
         }
-
-    }
-
-
-    [Authorize]
-    [ApiController]
-    [Route("api/[controller]")]
-    public class AdministrationController : Controller
-    {
-        private DataBaseContext user_data_context;
-        public AdministrationController(DataBaseContext user_data_context)
-        {
-            this.user_data_context = user_data_context;
-        }
-
-
-        // CREATE, EDIT, DELETE OPERATIONS
-        // INDEX
-        [Authorize(Roles = "Admin")]
-        [HttpGet]
-        public IEnumerable<Users> Get()
-        {
-            return user_data_context.User.ToList();
-        }
-
-        // DETAILS
-        [Authorize(Roles = "Admin")]
-        [HttpGet("{id}")]
-        public Users Get(int id)
-        {
-            return this.user_data_context.User.Where(user => user.Id == id).FirstOrDefault();
-        }
-
-        // CREATE
-        [Authorize(Roles = "Admin")]
-        [HttpPost]
-        public string Post([FromBody] Users New_User)
-        {
-            this.user_data_context.User.Add(New_User);
-            this.user_data_context.SaveChanges();
-            return "New_User created successfully!";
-        }
-
-
-        // EDIT
-        // [Authorize(Roles = Roles.Admin)]
-        [Authorize(Roles = "Admin")]
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Users New_User)
-        {
-            this.user_data_context.User.Update(New_User);
-            this.user_data_context.SaveChanges();
-        }
-
-        // DELETE
-        // [Authorize(Roles = Roles.Admin)]
-        [Authorize(Roles = "Admin")]
-        [HttpDelete("{id}")]
-        public void Delete(int id)
-        {
-            this.user_data_context.User.Remove(this.user_data_context.User.Where(New_User => New_User.Id == id).FirstOrDefault());
-            this.user_data_context.SaveChanges();
-        }
-
     }
 }
