+++ frontend/Controllers/ContactController.cs
@@ -62,9 +62,6 @@ namespace frontend.Controllers
 
 
 
-
-
-
         //INDEX
         [HttpGet]
         public async Task<IActionResult> Index()
@@ -88,99 +85,5 @@ namespace frontend.Controllers
             }
         }
 
-
-
-
-        // //EDIT
-        // [HttpGet]
-        // public async Task<IActionResult> Edit(int id)
-        // {
-        //     var response = await httpMsgClient.GetAsync(Configuration.GetValue<string>("WebAPIBaseUrl") + $"/contact/{id}");
-        //     response.EnsureSuccessStatusCode();
-        //     var content = await response.Content.ReadAsStringAsync();
-        //     var newMsg = new Contact();
-        //     if (response.Content.Headers.ContentType.MediaType == "application/json")
-        //     {
-        //         newMsg = JsonConvert.DeserializeObject<Contact>(content);
-        //     }
-        //     return View(newMsg);
-        // }
-
-        // [HttpPost]
-        // public async Task<IActionResult> Edit(Contact newMsg)
-        // {
-        //     if (ModelState.IsValid)
-        //     {
-        //         //httpMsgClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-        //         httpMsgClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", HttpContext.Session.GetString("token"));
-        //         var serializedProductToEdit = JsonConvert.SerializeObject(newMsg);
-        //         var request = new HttpRequestMessage(HttpMethod.Put, Configuration.GetValue<string>("WebAPIBaseUrl") + $"/contact/{newMsg.Id}");
-        //         request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-        //         request.Content = new StringContent(serializedProductToEdit);
-        //         request.Content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
-        //         var response = await httpMsgClient.SendAsync(request);
-        //         if (response.IsSuccessStatusCode)
-        //         {
-        //             return RedirectToAction("SentMessage", "Messages");
-        //         }
-        //         else
-        //         {
-        //             // ViewBag.Message = "Admin access required";
-        //             // return View("Edit");
-        //             return RedirectToAction("Error401", "Error");
-        //         }
-        //     }
-        //     else
-        //         return RedirectToAction("Error401", "Error");
-        // }
-
-
-
-
-
-        // //DELETE
-        // [HttpGet]
-        // public async Task<IActionResult> Delete(int id)
-        // {
-        //     var response = await httpMsgClient.GetAsync(Configuration.GetValue<string>("WebAPIBaseUrl") + $"/contact/{id}");
-        //     response.EnsureSuccessStatusCode();
-        //     var content = await response.Content.ReadAsStringAsync();
-        //     var contactMsg = new Contact();
-        //     if (response.Content.Headers.ContentType.MediaType == "application/json")
-        //     {
-        //         contactMsg = JsonConvert.DeserializeObject<Contact>(content);
-        //     }
-        //     return View(contactMsg);
-        // }
-
-        // [HttpPost]
-        // public async Task<IActionResult> Delete(Contact contactMsg)
-        // {
-        //     if (ModelState.IsValid)
-        //     {
-        //         //httpMsgClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-        //         httpMsgClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", HttpContext.Session.GetString("token"));
-        //         var serializedProductToDelete = JsonConvert.SerializeObject(contactMsg);
-        //         var request = new HttpRequestMessage(HttpMethod.Delete, Configuration.GetValue<string>("WebAPIBaseUrl") + $"/contact/{contactMsg.Id}");
-        //         request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-        //         request.Content = new StringContent(serializedProductToDelete);
-        //         request.Content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
-        //         var response = await httpMsgClient.SendAsync(request);
-
-        //         if (response.IsSuccessStatusCode)
-        //         {
-        //             return RedirectToAction("SentMessage", "Messages");
-        //         }
-        //         else
-        //         {
-        //             // ViewBag.Message = "Admin access required";
-        //             // return View("Delete");
-        //             return RedirectToAction("Error401", "Error");
-        //         }
-        //     }
-        //     else
-        //         ViewBag.Message = "wrong";
-        //     return View("Delete");
-        // }
     }
 }
